Add files via upload
Passes the test cases by properly figuring out the safety of the case. However, the printing path should be arranged. In the algortihm, once an executable process is found, it iterates again from the first iteration, whereas it should continue its active path and get back to the start when it comes to the end.